Backend/requirements.txt
argon2-cffi
asgiref
channels[daphne]
channels-redis
cryptography
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
gunicorn
psycopg2-binary
PyJWT
pyotp
python-dotenv
pytz
qrcode
sqlparse